PythonServer/requirements.txt
numpy
torch
keras
pillow
keras2onnx
onnxruntime
h5py